disp: msm: dsi: add a lock free API to enable DSI core clks
In dual display cases, the calls to enter and exit idle cases
already happen from the context of the dsi_display_clk_ctrl() API.

When the slave display needs to turn on the master DSI core clks
it doesn't need to again go through the locked version as it results
in a deadlock.

Introduce a lock free version of dsi_display_clk_ctrl() to address
this use-case.

Change-Id: I615095d8c4f7b09a34b049342be081f92a6d3e3a
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>
Signed-off-by: Srihitha Tangudu <quic_tangudu@quicinc.com>