disp: msm: dsi: disable master PHY only when the display is off
When the slave PHY turns off as part of display OFF, master PHY needs
to be disabled in case that the master display is not being used. Currently
master PHY's ref count is checked in order to decide if the display is ON
or OFF. This is an incomplete check as the ref count can become zero in
case of master display going into idle. The more accurate check therefore
will be to see if the display panel is initialized or not.

The change also adds a more verbose event log at the start of the PHY
enable/disable functions rather than at the end.

Change-Id: I96043919b5c219b92ba7938e4606e7fbb6b3b242
Signed-off-by: Satya Rama Aditya Pinapala <psraditya30@codeaurora.org>
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>
Signed-off-by: Srihitha Tangudu <quic_tangudu@quicinc.com>