mm-drivers: hw_fence: add hw-fence refcount for create and signal
This change adds refcounting to the hw-fence for creation and signaling.
When a hw-fence is created, it has two refcounts, which are released by
the hw-fence client that created the fence and the fence controller when
it signals the fence. The hw-fence is destroyed when the hw-fence
refcount reaches zero.

Change-Id: I812bf8ce7a35df22390d0e77f10c0797152d57a3
Signed-off-by: Grace An <quic_gracan@quicinc.com>