mm-drivers: hw_fence: add dma-fence signal callback
This change adds a dma-fence signal callback for dma-fences backed
by hw-fences. To ensure that the hw-fence is not destroyed until the
dma-fence is signaled, a refcount on the hw-fence is stored for the
dma-fence signal callback. When the dma-fence is signaled, the hw-fence
refcount is decremented by one. If the hw-fence has not been signaled,
signal the hw-fence from the HW Fence Driver.

Change-Id: I0d511f9c9c64f12ea2affe63fb221de2a267331e
Signed-off-by: Grace An <quic_gracan@quicinc.com>