dsp-kernel: detect privileged processes based on group ID
Get the information on privileged group IDs during rpmsg probing based
on DT property. Check if the process requesting an offload to remote
subsystem is privileged by comparing it's group ID with privileged
group ID. Initialization process attributes are updated for a
privileged process which is sent to remote process for resource
management.

Signen-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>