dsp-kernel: Add cached and persistent buffer support
FastRPC driver uses DMA API for region allocation for any buf_alloc
request and the DMA region is unmap and freed once the usage is
complete.

Cached buffer support enables caching of certain types of buffers
instead of freeing which help in reducing the overhead of DMA API
call for every buffer allocation request.

A list of cached buffer is maintained which will get reused when
needed and this buffer list will get freed during device release.

This change also enables the usage of persistent header buffers which
are allocated during process initialization and is mapped to CDSP.
This mechanism helps in reducing the RPC overhead for remote calls.

For any user process remote RPC call to CDSP, it is tried to use
persistent header buffers. In case all persistent buffers are
already in use, there is a fallback to cached buffer usage and
allocation using DMA API.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: Id86cdd515e590108c6a47750fe58086e93895d45