disp: msm: sde: expose ubwc-clk property in crtc
Add ubwc-clk to CRTC property list to give usermode the ability
to set it separately. Currently ubwc-clk calculation is based on
all active displays and finally the max of ubwc-clk & core-clk is
used to get the final core-clk. With Display CESTA, since the
client aggregation is moved to HW, ubwc-clk property gives the
ability to program core-clk & ubwc-clk separately and let HW do
the aggregation & max calculations between clients.

Change-Id: Iab9fa0e9727ab5c45eba6cdfd762ea8b585ca722
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>