disp: msm: sde: reduce regbus vote in power on commit
Current SDE driver keeps max vote in power on commit along
with idle-pc restore. In power on commit, since duration is
long because of panel on, regbus scaling to max may not be
required. This change reduces regbus vote only on power on
commit and keeping max vote in idle-pc restore case as is.

Change-Id: I1ea11bf8de145870a7692415a581e29349a75410
Signed-off-by: Prabhanjan Kandula <quic_pkandula@quicinc.com>