mm-drivers: hw_fence: use different lock implementation for soccp
On targets with fencing workload running on a secondary vm, the inter-vm
lock uses two bits to distinguish lock owner between hw-fence driver and
vm. If the vm fails to acquire an inter-vm lock, then it will go to sleep.
Upon unlocking the contended inter-vm lock, the hlos must wake up the vm
via an ipcc signal.

This change adds lock implementation for targets with soccp: For inter-
processor lock with soccp, only one bit is needed for lock and unlock,
and ipcc signal should not be triggered for unlock.

Change-Id: I910dbe16237196621a25fe9e70bfd2d4bbce5674
Signed-off-by: Grace An <quic_gracan@quicinc.com>