disp: msm: use GenPD instead of regulators for MDSS GDSC
From Sun target, clock driver moved the GDSC from regulator framework to
GenPD to match with the upstream. Add support to use the GenPD using
pm_runtime apis. With RSCC, MDSS GDSC is in HW-controlled mode and the
setting is done as part of the initial pm_runtime_get_sync as opposed
to a set_mode call with regulator framework. For targets without RSCC,
adding the power-domain DT entry to the msm_mdss node should be
sufficient as the existing pm_runtime usage for tracking clk/bw/regulator
resources should be sufficient to control the regulator.

Change-Id: Ib5575aa8d061ffb48f0114518c733cf8eb9d196c
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>