dsp-kernel: Free allocated perf object instead of user pointer
If perf counters tracing is enabled, a perf object is allocated during
context alloc and this object is expected to be freed during context
free. As of now, there is a problem due to which free for user pointer
perf_kernel is getting called causing page fault. Add changes to
fix this issue.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: I7d48ce67518aa3aba6577d6920b3c2c2d51ea848