dsp-kernel: Support DMA handle in no argument case
DSP updates fd of DMA handles to be freed in fdlist. These
FDs are then expected to be freed during put_args. The allocation
for fdlist happens in get_args and the map corresponding to the
FD will be freed in put_args. This list can also be updated over
a call with no input or output arguments.

Signen-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: I3a3a22b8e33ec2e0a08af89b42f05d36aee2d143