disp: msm: use GenPD instead of regulators for MDSS GDSC with CESTA
From sun target, clock driver moved the GDSC from regulator framework
to GenPd to match with upstream. Add support to use GenPD using
pm_runtime apis. With Display CESTA, MDSS GDSC is in HW-controlled
mode and the setting is done as part of the initial pm_runtime_get_sync
as opposed to a set_mode call with the regulator framework.

Change-Id: Iee2b7f972b8645179cd0ce51db0e7bd3ab63a0a6
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>