dsp-kernel: Add support to allocate shared context bank
Context banks could be set as a shared one using a DT propery
"qcom,nsessions". The property takes the number of session to
be created of the context bank. This change provides a control
mechanism for user to use shared context banks for light weight
processes. The session is set as shared while its creation and
if a user requests for shared context bank, the same will be
allocated during process initialization.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: I0b5392eec2e042b10e96d27a54b4e3069248ed81