dsp-kernel: Skip reference for DMA handles
If multiple dma handles are passed with same fd over a remote call
the kernel driver takes a reference and expects that put for the
map will be called as many times to free the map. But DSP only
updates the fd one time in the fd list when the DSP refcount
goes to zero and hence kernel make put call only once for the
fd. This can cause SMMU fault issue as the same fd can be used
in future for some other call.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: Idb63a66eb8fe82a28af5c0ad4da181856b0ce98d