dsp-kernel: Add va and length check when looking for map
If any fd is mapped with some va and size and a mapping request comes
with a large size with same fd, the lookup function end up
incrementing the refcount of old map without checking the size. The
same map is then passed to DSP which might end out access memory
which is not mapped by kernel, hence resulting in memory access
faults. Add va and length when checking for maps to take a refcount
for already existing maps.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: Idbae78a15c00b5ed560c0f6b05c04e6415a2edd8