dsp-kernel: Static PD restart support and remote heap management redesign
Static PDs need to support restart feature where it can be restarted
affecting other PDs running on the same subsystem. For this support
fastrpc driver needs to handle events of PD down and up. This
requires addition of a service locator to get the event notifications
for audio and sensor domain static PDs. On different events, fastrpc
needs to take care of cleaning up resources and notifying the user
daemon process.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: I79ea02807692a1ed120f0882824b070671c79dba