disp: msm: get correct dsi display info for shd connectors
In the case of SHD display, if the connector type is shared,
then assign it to the dsi_base of shd display else typecast it
to the dsi_display structure. This will help to avoid the out of memory
issues, as dsi_display is a pointer in shd_display structure and assigning
it directly to dsi_display structure will cause out of memory issues.

Additionally, add encoder vblank locks during register/unregister of
the IRQ to protect it from multiple threads. Allocate DS block during
allocation of resources in shared display.

Change-Id: I7d4b64e69087dccbf506c85d89ecf75f1103772a
Signed-off-by: Rajesh kv <quic_kvrajesh@quicinc.com>