asoc: lpass-cdc-va-macro: revert reset Sampling rate
Sampling rate of the decimator is reset to hw reset value during
power down event in existing code. This code was implemented for
 a safety precaution.
Remove this code as it is causing a side effect for VA use cases.

Side effect of existing code: during successive VA use cases runs,
Sampling rate is not updated as the device is not closed completely.
Due to this, decimator runs with incorrect sampling rate for
 successive runs and causes failure.

Change-Id: I5e60c656043d06c3d999de183698adfffdaa44d1
Signed-off-by: Vangala, Amarnath <quic_avangala@quicinc.com>
(cherry picked from commit daea160320cc258ae01eefa5d74cf3721a3fe0cc)