touch: sun: fix touch abnormal issue for sun target
There are two issues observed, one issue is pointer
location not working correctly, can only see white
point when moving finger on panel, by removing
unnecessary input ability can fix this issue.
The other issue is crash logs is observerd, it is
because when process is holding a spinlock and call
disable_irq, it might sleep which is not allowed,
changing to use mutex can fix this issue.

Change-Id: I590fdd051d07afc32f21b4f1c5b731cde5535e62
Signed-off-by: Yu Wu <quic_zwy@quicinc.com>