disp: msm: sde: assign a non-pairable DSC to a topology that requires single DSC
During suspend-resume in a dual-display scenario with primary
display topology <2 2 1> and secondary display topology <1 1 1>
with 3 DSC instances, allocating DSC block 1 to the secondary
display results in DSC blocks 2 and 3 being assigned to the
primary display. However, as DSC block 2 is not a peer of DSC
block 3, it triggers an atomic check failure as a part of the RM
reservation failure. While reserving the DSC resources, prefer non
pairable DSC if the topology requests for single DSC instance. If
there are no non pairable DSCs available, then select one of the
available DSC.

Change-Id: I3e3f7ec343b5bc99aa23786a3cb65cf7bbbd7275
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>