dsp-kernel: Restrict untrusted applications from attaching to GuestOS
Untrusted application can attach to guestOS and staticPD if it can
somehow make INIT IOCTL call with ATTACH flag. This is a potential
security issue as the untrusted application can crash guestOS or
staticPD. Restrict attach to guestOS or staticPD request if request
is being made using non-secure device node.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Change-Id: I5149eca82882b0c870ea7a26ac06208a9c95e90a