dsp-kernel: Check for untrusted process during process init
Untrusted application is not allowed to offload to system unsigned
PD. The untrusted apk check is part of set_session_info which can
be skipped by the client to directly call init_process. Add a check
for untrusted process in init_process request to make sure that
system unsigned is not allowed for untrusted apk if the client skips
set_session_info call.

Change-Id: I0b99627206666b40233226cba699ab0a5c9b1360
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>