qcacld-3.0: Enable WOW_BEACON_EVENT for STA mode
Today in WOW mode, if the beacon is received and interested
IE changes, FW is checking whether the WOW_BEACON_EVENT flag
is enabled or not to allow the beacon to be indicated to HOST
and trigger a WOW wakeup.
So, enable WOW_BEACON_EVENT for STA mode and p2p client mode to get
such beacon frames in WOW mode.
And use correct the BSS color change announce IE ext id in beacon
filter.

Change-Id: I37f0c5cac5b8324fe50fde5f4bef06b4a0d15267
CRs-Fixed: 3703517