disp: msm: sde: add support to get active LM data from CTL path
In video mode with TUI use case, a null commit is executed
before transitioning to the secondary VM. This causes the reset
of the number of active LMs on Primary VM. In this scenario
atomic check in the secondary VM fails because of no availability
of appropriate mixers. Retrieve active LM info from MDSS HW
to address this issue.

Change-Id: I67d1877bb768c3a486d5d655e085b4e730657395
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>