msm: camera: isp: Prevent OOB on accessing path irq desc array
Currently, the irq status is right shifted one bit at a time
until all the error or info bit have been logged. But, the bit
position is not bounded to how bits are actually supported and
mentioned in the header. This could cause an out of bounds access.
Validate the bit position based on number of bits supported in
the irq status register specified in header. Also, address the
gap in v980 header which triggered this issue.

CRs-Fixed: 3726411
Change-Id: Ia6c72aab496bd27d0c2d3ecbfa9999014b9403fe
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>