dsp-kernel: Avoid incorrect buffer deletion
Static process init request allocates 2MB buffer for dynamic loading
and this buffer is only freed for SSR and PDR case. After first
init request all other init requests should skip this buffer
allocation. With current design, the allocation is getting skipped
but there is an attempt to delete invalid buffer on failure. Also
the staticpd_status is getting reset for failure which might result
in memory leak if the buffer was not allocated for the init request.
Add changes to delete buffer and set staticpd_status only if the
buffer was allocated during init call.

Change-Id: Id408fbd2a74a527ae2f885ce2155359a24bcfe7b
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>