msm: camera: smmu: Remove additional mapping refcount
Camera SMMU can support multiple handles for the same context
bank. When mapping a buffer from a given bank to these handles,
we internally refcount the buffer for each handle. On buffer release,
it is unmapped as a whole for the given bank. But due to the multiple
ref counts, unmap to SMMU is not invoked correctly. This change avoids
adding ref count for different handles in the same bank.

CRs-Fixed: 3713983
Change-Id: I89c554af9327d4edf800c6e6c268ffaf30b77f82
Signed-off-by: Karthik Anantha Ram <quic_kartanan@quicinc.com>