mm-drivers: hw_fence: support import by creating client
Currently, if a client that created the fence calls synx_import, the
client registers to wait on the fence.

This change ensures that synx_import by the fence allocator does not
result in this client getting notified of fence signal. Instead this
function increments hlos refcount and sets the fctl refcount if not
already set for the hw-fence.

Change-Id: I55772a46cf7f428d43be02b9f3e87102e9f2d54a
Signed-off-by: Grace An <quic_gracan@quicinc.com>