msm: synx: Resolving derefencing of synx_cb after free
There is a race condition between timer expiry and signal of handle in
which one of them queues the work for cb dispatch and other thread tries
to access synx_cb after it gets free. This change makes sure that we
lock protect the reg_cb_lists and check if the cb node is present in
list before dispatching the work.
Since mutex lock cannot be taken in atomic context for timer callback,
the operation is pushed to worker thread.

Change-Id: I4bd3b7e4e87ac46267e6a9c288d66622175d0bbb
Signed-off-by: Amir Suhail <quic_asuhail@quicinc.com>