msm: camera: icp: Change SSR sequence for ICP
Currently, when ICP encounters a fatal error, HLOS resets the
HW cores that ICP FW handles. Reload and reset the ICP core in the
next camera session. To prevent HLOS from resetting the HW cores,
change the sequence to reload and reset ICP in the failing
camera session. ICP FW would then reset the cores it handles.
With this sequence change there is no need for HLOS driver
to access IPE/BPS or OFE cores.

CRs-Fixed: 3749708
Change-Id: I4be07a1e952b0af0e2db4978107b573060cee922
Signed-off-by: Karthik Anantha Ram <quic_kartanan@quicinc.com>