disp: msm: add esync support
Adds support for the esync hardware block.

Adds the esync and oscillator clocks that drive the main and
backup esync generators, respectively. Controls these clocks such
that esync clock for a given interface is running only when its
esync generator is running, and the same for the oscillator clock.

Programs and enables the main esync generator when in active usecase,
and programs and enables the backup esync generator when
transitioning to idle.

Adds the means of sending panel specific commands at timing
available from SDE driver, such as to send commands right after
enabling the esync generator when resuming.

Change-Id: I09897cdb7220c5909fae43e2bc9455f6f28bdb7c
Signed-off-by: Kirill Shpin <quic_kshpin@quicinc.com>