msm: camera: common: Sanitize event payload before adding to free list
Currently, during irq handling, we are putting the event payload to a
free list in the bottom half handler, when we are done using it. But,
there is data in the list entry which is not reset before putting the
entry to free list, which might cause issues when using the same entry
in subsequent irq handling. Fix this issue using a macro which sanitizes
the list entry data before putting the payload to free list.

CRs-Fixed: 3747093
Change-Id: I0bd0f16585097825063701102697da4ec94bcc31
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>