disp: msm: dp: fix DSC byte padding and alignment for 4HS merge topology
While passing input bytes to the DCE in 4 hard slice mode, programming
should make sure that the number of bytes per DSC pair is aligned to 8.
If not, byte padding and alignment should be added. In the current
implementation, formula to calculate the total bytes in a DSC pair
was wrong leading to uneven byte padding and alignment for specific
resolutions, such as 6016x3384. This change fixes this calculation and
also sets the default byte padding and alignment value.

Change-Id: Ib5c4a69ea8f790ad523db64e87430a2744b7bbbd
Signed-off-by: Sandeep Gangadharaiah <quic_sandgang@quicinc.com>