msm: camera: isp: Adding null check before freeing FCG struct
This change adds a null check to fix a kernel panic issue
reported when trying to dereference a null variable. In this case,
a call to acquire device is made then followed by release device which
calls an api that ultimatley frees fcg caps. These are only allocated by
a call to acquire hardware which was not made in this scenario thus fcg
caps are null and dereferencing them caused this issue.

CRs-Fixed: 3760279
Change-Id: I5c81ed0baf0e5d838e48e07b660b8dd4bdcedd63
Signed-off-by: Atiya Kailany <quic_akailany@quicinc.com>