qseecom: update dma_parms memory allocation style
In kernel version 5.15 and newer, the dev.dma_parms is filled before
platform driver probe, hence it is potentially harmful to kfree it in
deinit.

This commit changes the allocation style from kzalloc to devm_kzalloc,
the memory being allocated will be bound to device, and will get auto
cleaned when driver destroy, without the neccessity to explicitly call
kfree. In this way we do not need to care about if dma_parms is filled
previously by the kernel or not.

Change-Id: Ica6617240cf357a609132086b42cdd72349dad4c
Signed-off-by: Yuhang Ji <quic_yuhaji@quicinc.com>