dsp-kernel: Add process session check in dmabuf_alloc
FastRPC provides an interface for user to allocate memory from
the process context bank session. With the new design, the
session allocation operation is moved from device_open to
process_init. This might create problems if the user is
requesting for buffer allocation before process_init call.
Add changes to gracefully handle this scenario and return
failure if session is not allocated when user requests for
buffer.

Change-Id: Ifc07ef4feddd08a39b80aa14f7abb7ac0c1ac082
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>