disp: msm: sde: avoid cesta override setting for 2 frames
With display cesta, the cesta ctl flush needs to be set with
override vote for the first frame as the vote needs to be applied
immediately. Avoid setting it for the next frame as it may lead
to dpu hang if its updated while a transfer is in progress as the
idle will assert at the end of frame transfer and there is no flush
to add the next vote with auto-vote-on-idle disabled.

Change-Id: I5c0a187b8006b0e243e0d389e2b214b737449825
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>