icnss2: enlarge maximum length for filtering rproc name
The remote proc name for wpss on volcano is
'9bb00000.remoteproc-wpss', the restart level changing
on which will be ignored by icnss platform driver
unexpectedly due to the maximum length for filtering the
name is too short.

To fix it, enlarge this maximum length to 100 bytes
according to implementation of remote proc.

Change-Id: I42148cb2f8d0821521acf06e5f54d4ebaf439985
CRs-Fixed: 3794773