dsp-kernel: Avoid duplicate mapping of shell memory
If fastrpc shell is opened by the application, it's content is read
to a buffer and it is passed to the remoteproc during PD init call.
As part of PD init call, the buffer is getting mapped to SMMU twice
which is not required. First the map is created in init_create_process
if the file fd is passed from the user space. Then this file fd is
added to the invoke arguments to be passed to the remoteproc. This
would cause a increament in the ref count of the map during get_args.
Add changes to avoid duplicate mapping of shell memory.

Change-Id: Idc2e52278048e868f9e8fd0458f2e4c3bca9d169
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>