dsp-kernel: Make gmsg_log a single struct in channel context
Channel context carries glink transaction logs as array of struct
which is not required as gmsg_log structure is unique to channel.
Add changes to remove gmsg_log as array member and have it as a
single struct in channel context struct. Also add minor changes
to queue session info also for SSR notification and change to
handle rpmsg_probe failure properly.

Change-Id: I86272ddc9965aed38f7525d8d32b63a36ad34d93
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>