msm: camera: icp: Improve message process for HFI queue
Going by current scheme, even if the message size is larger than
the size of drain buffer, it will only read once from HFI queue and
thus fail to read. This change eliminates the requirement that the
size of drain buffer by reading the msg or dbg queue multiple times
until the data in the queue is empty.

CRs-Fixed: 3591653
Change-Id: I60008166f32e3019e244a38611b23830c160bc68
Signed-off-by: Stark Lin <quic_starlin@quicinc.com>