disp: msm: sde: move parsing of soccp dt to init power resources
Currently, the display driver parses the soccp device-tree nodes
as part of initializing the hw blocks. This happens after display
sets the power votes for the first time, preventing display from
setting the power vote for soccp for this instance. Thus, move
parsing of soccp dt to initialization of power resources so that
display correctly sets the power vote for the first power enable.

Change-Id: Iae4a1bed4972e6367d70cdf8f4674d56a36198e6
Signed-off-by: Grace An <quic_gracan@quicinc.com>