dsp-kernel: Use msg PID to identify which userPD to kill on the DSP
Currently, arguments are mapped on the DSP to obtain the PID for
identifying which userPD to terminate on the DSP. If the DSP runs out
of resources, such as VTLB entries, the process kill request will not
be serviced, resulting in a SMMU fault. Pack the PID in the message at
HLOS and use the same to kill the process on DSP so that process kill
will always be serviced regardless of whether the DSP has run out of
resources or not.

Change-Id: Ifd5ab6785ad6dd369acbd21c33a2bc0aedce3f07
Signed-off-by: Abhishek Singh <quic_abhishes@quicinc.com>