dsp-kernel: Capture boot time in gmsg_logs
We capture details for all the glink transactions and one of the
details captured is the time of the message rx or tx. As of now
the real time is captured in the gmsg_logs which might not be
useful. Capturing boot time might help in debugging more
efficiently as the gmsg_logs can be matched with the failure
logs. Add changes to capture boot time instead of real time.

Change-Id: Iafe2572646c229ae85ed6af3e3dea2e75520b899
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>