disp: msm: sde: allocate memory with kvzalloc for rc block
The kzalloc family of APIs always allocates both virtually
contiguous and physically contiguous memory. In a TVM, sometimes
scarcity of physically contiguous memory can lead to memory
allocation failures, triggering an out-of-memory condition and
causing the TVM kernel to crash. To avoid this scenario, allocate
memory using the kvzalloc API.

Change-Id: Ia29477eb0635788a2b4ea0d03fa1a04f325b085c
Signed-off-by: Akash Gajjar <quic_agajjar@quicinc.com>