dsp-kernel: Add a different state for remote processes
User can call map and unmap IOCTLs before the process init IOCTL.
Session and device information are necessary for creating and destroying
the map, which is allocated as part of process init. This approach can
leads to a NULL dereference scenarios. To address this, add a different
state for remote processes and check it before proceeding with mapping
or unmapping.

Change-Id: I5764c01f624547959b64d46cb72e404f192fda77
Signed-off-by: Santosh <quic_ssakore@quicinc.com>