dsp-kernel: release all pdr handles in case of ssr
Currently, the pdr handle for ois pd on adsp is not being released in
case of an ssr, resulting in a leak. Release all pdr handles on ssr.

Also, while searching for matching clients for a driver handle, replace
redundant code to loop thru channel contexts with a for loop.

Change-Id: I9c7d7d2e7b65c2e71bd4d5e45f5f6ed1b99d14e6
Signed-off-by: Thyagarajan Venkatanarayanan <quic_venkatan@quicinc.com>