msm: camera: isp: Avoid OOB access of deferred buf-done indices
In SHDR use-case, We see array out of bound access as
differ buffer-done exceeding ISP max number of resources.
We were handling the BUF_DONE event for a requests that
is still in wait-list, do not signal now, instead mark
it as done and handle it later.
Placeholder for holding deferred buf_done indices is
static value which can be lesser then the number of
ports acquired based on the use-case or target. In corner
cases the index can overshoot the static memory.
This commit moves the static memory to dynamic allocation
to take care of number of ports acquired. This takes care
of use-cases and number of clients changing per target.

CRs-Fixed: 3738424
Change-Id: Ic0710320e3af9cafc807fbbc168407c1c99dca39
Signed-off-by: Chandan Kumar Jha <quic_cjha@quicinc.com>