msm: camera: isp: Add support for 4DTs per VC
In v1080, hw support is expanded such that each
VC can be matched with up to 4 DT. On top of
which previously, VC0 can only be tied to DT0,
and VC1 to DT1, in v1080, any VC can be tied to
any 4 DTs.

This change is added to support this HW
enhancement.

CRs-Fixed: 3783559
Change-Id: I637774134e0e90ffb8811b6d138667a0078543df
Signed-off-by: Li Sha Lim <quic_lishlim@quicinc.com>