msm: synx: add synx signaling status for external ssr errors
HW Fence Driver may invoke signal_fence both when SOCCP and HW Fence
clients undergo SSR. When SOCCP is undergoing SSR, the Synx Driver
must unlock the synx global object in case it was locked prior to
crash. If SOCCP is not undergoing SSR, the Synx Driver should not
release the lock of the synx global object.

The is_core_ssr bool input parameter indicates the first scenario
where the Synx Driver should release the lock of the synx global
object. If this is false, the Synx Driver will not release the lock
of the synx global object.

Change-Id: I1135c8079ff74a884cd8bb75131c64af716f5281
Signed-off-by: Grace An <quic_gracan@quicinc.com>