qcacmn: Fix spin lock during ctrl pkt and filter hdl list operation
This change is to fix spin unlock during control packet and filter
handle enqueue in the list for exceptional case where the list
is completely exhausted.

The list is used in work queue handler where every entry is dequeued
and processed. It also applies spin lock while dequeuing the entry
from the list to avoid overwrite of hp and tp value from different context.

Change-Id: Ie5b4b31c0b7ed6b31dc14eba6055aaaf79845f6d
CRs-Fixed: 3825184