disp: msm: dp: increase wait time for HDCP polling mode state transition
During the HDCP AKE stage, if a CP-IRQ is received right after the
successful completion of the HDCP authentication part 2, the IRQ handler
gets invoked even before the HDCP state engine gets transitioned
to the polling mode. This can cause the test to fail as we would not
read the RepeaterAuth_Send_ReceiverID_List from the TE in response to
the CP_IRQ. Hence, currently driver waits for 300 msecs in CP_IRQ attn
work for processing this request. However, in latest releases the time
taken for HDCP engine to transition into polling mode has increased to
approx. 500 msec. This change increases this waiting time from 300 to
600 msecs. This change also adds a missing debug print during HDCP state
transitions.

Change-Id: If247965231e8b172e5edcd21f51c4fcc9f5311c9
Signed-off-by: Sandeep Gangadharaiah <quic_sandgang@quicinc.com>