disp: msm: sde: force modeset during cac loopback transitions
Entry and exit of cac loopback cases is a seamless transition
but it needs modeset as resource allocation and setting up
of CTL path will change based on whether loopback is enabled
or not. This change adds support to detect loopback transition
cases and force encoder modeset to make sure that reservations
and pipeline is setup properly.

Change-Id: I50ec66c8dff52c6199ae7a5dcc44a73d9c80d667
Signed-off-by: Krishna Manikandan <quic_mkrishn@quicinc.com>
Signed-off-by: Sampurna Bolloju <quic_sampboll@quicinc.com>