disp: msm: sde: issue cesta scc reset during ctl-reset
Currently ctl reset is issued during esd-failures or dpu hang issues.
Add cesta scc reset to this sequence when cesta is enabled to reset
the cesta states to recover fully.
As part of the change, remove the reset triggered from frame-done
timeout, so that the complete reset sequence can be executed during
the next frame update.

Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>
Signed-off-by: Xhoendi Collaku <xcollaku@quicinc.com>
Change-Id: Iae728a6cdb7987b9797c23d89e47012ae9a59902
Signed-off-by: Xhoendi Collaku <quic_xcollaku@quicinc.com>