disp: msm: sde: issue cesta scc reset during ctl-reset
Currently ctl reset is issued during esd-failures or dpu hang issues.
Add cesta scc reset to this sequence when cesta is enabled to reset
the cesta states to recover fully.
As part of the change, remove the reset triggered from frame-done
timeout, so that the complete reset sequence can be executed during
the next frame update.

Change-Id: If94a2cdff3e1997d9ea831477dfa97b4a37a9037
Signed-off-by: Veera Sundaram Sankaran <quic_veeras@quicinc.com>